explain why [...docs].jsx, merge with documentation.jsx
The reason I removed it is because having two files that refer to
documentation, i.e., "[...docs].jsx" and "documentation.jsx" is
confusing.
